refactor(inventory): extract bundle helper types into dto package
