+++ src/main/java/com/spotcamp/module/inventory/controller/BundleController.java
@@ -1,6 +1,8 @@
 package com.spotcamp.module.inventory.controller;
 
 import com.spotcamp.module.inventory.dto.BundleDefinitionDTO;
+import com.spotcamp.module.inventory.dto.BundleAvailabilityResult;
+import com.spotcamp.module.inventory.dto.BundleComponentRequest;
 import com.spotcamp.module.inventory.dto.BundleRequestDTO;
 import com.spotcamp.module.inventory.entity.Bundle;
 import com.spotcamp.module.inventory.service.BundleService;
@@ -104,9 +106,9 @@ public class BundleController {
     ) {
         log.info("Creating new bundle: {}", request.getName());
         
-        List<BundleService.BundleComponentRequest> components = request.getComponents().stream()
+        List<BundleComponentRequest> components = request.getComponents().stream()
                 .map(comp -> {
-                    BundleService.BundleComponentRequest componentReq = new BundleService.BundleComponentRequest();
+                    BundleComponentRequest componentReq = new BundleComponentRequest();
                     componentReq.setProductId(comp.getProductId());
                     componentReq.setQuantity(comp.getQuantity());
                     return componentReq;
@@ -192,9 +194,9 @@ public class BundleController {
     ) {
         log.info("Updating bundle: {}", bundleId);
         
-        List<BundleService.BundleComponentRequest> components = request.getComponents().stream()
+        List<BundleComponentRequest> components = request.getComponents().stream()
                 .map(comp -> {
-                    BundleService.BundleComponentRequest componentReq = new BundleService.BundleComponentRequest();
+                    BundleComponentRequest componentReq = new BundleComponentRequest();
                     componentReq.setProductId(comp.getProductId());
                     componentReq.setQuantity(comp.getQuantity());
                     return componentReq;
@@ -276,7 +278,7 @@ public class BundleController {
     ) {
         log.debug("Checking availability for bundle: {} from {} to {}", bundleId, checkIn, checkOut);
         
-        BundleService.BundleAvailabilityResult result = bundleService.checkBundleAvailability(
+        BundleAvailabilityResult result = bundleService.checkBundleAvailability(
                 bundleId, checkIn, checkOut);
         
         Map<String, Object> response = Map.of(
