+++ src/main/java/com/spotcamp/module/inventory/dto/BundleComponentRequest.java
@@ -1,0 +1,22 @@
+package com.spotcamp.module.inventory.dto;
+
+public class BundleComponentRequest {
+    private Long productId;
+    private Integer quantity;
+
+    public Long getProductId() {
+        return productId;
+    }
+
+    public void setProductId(Long productId) {
+        this.productId = productId;
+    }
+
+    public Integer getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(Integer quantity) {
+        this.quantity = quantity;
+    }
+}
